requirements.txt
# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Web framework
flask==3.0.0
flask-cors==4.0.0

# LLM integration
openai==1.3.0
anthropic==0.7.8

# Data processing
pandas==2.1.3
json5==0.9.14

# Caching
redis==5.0.1
diskcache==5.6.3

# PDF processing  
PyPDF2==3.0.1
pdfplumber==0.10.3

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
typing-extensions==4.8.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0